requirements.txt
Flask
requests
python-dotenv
gunicorn
discord.py